requirements.txt
eventlet==0.18.4
Flask==0.10.1
Flask-SocketIO==2.1
Flask-SQLAlchemy==2.1
gevent==1.0.2
gevent-websocket==0.9.5
greenlet==0.4.9
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pyserial==3.0.1
python-engineio==0.8.8
python-socketio==1.0
six==1.10.0
SQLAlchemy==1.0.12
Werkzeug==0.11.4
wheel==0.29.0
